hopr_utils::runtime::prelude

Struct RwLockWriteGuard

pub struct RwLockWriteGuard<'a, T>
where
    T: ?Sized,
{
    pub(super) permits_acquired: u32,
    pub(super) s: &'a Semaphore,
    pub(super) data: *mut T,
    pub(super) marker: PhantomData<&'a mut T>,
}

RAII structure used to release the exclusive write access of a lock when dropped.

This structure is created by the write method on RwLock.

Fields
permits_acquired: u32s: &'a Semaphoredata: *mut Tmarker: PhantomData<&'a mut T>
Implementations

impl<'a, T> RwLockWriteGuard<'a, T>
where T: ?Sized,

pub fn map<F, U>( this: RwLockWriteGuard<'a, T>, f: F, ) -> RwLockMappedWriteGuard<'a, U>
where F: FnOnce(&mut T) -> &mut U, U: ?Sized,

Makes a new RwLockMappedWriteGuard for a component of the locked data.

This operation cannot fail as the RwLockWriteGuard passed in already locked the data.

This is an associated function that needs to be used as RwLockWriteGuard::map(..). A method would interfere with methods of the same name on the contents of the locked data.

This is an asynchronous version of RwLockWriteGuard::map from the parking_lot crate.

Examples

use tokio::sync::{RwLock, RwLockWriteGuard};

#[derive(Debug, Clone, Copy, PartialEq, Eq)]
struct Foo(u32);

let lock = RwLock::new(Foo(1));

{
    let mut mapped = RwLockWriteGuard::map(lock.write().await, |f| &mut f.0);
    *mapped = 2;
}

assert_eq!(Foo(2), *lock.read().await);

pub fn downgrade_map<F, U>( this: RwLockWriteGuard<'a, T>, f: F, ) -> RwLockReadGuard<'a, U>
where F: FnOnce(&T) -> &U, U: ?Sized,

Makes a new RwLockReadGuard for a component of the locked data.

This operation cannot fail as the RwLockWriteGuard passed in already locked the data.

This is an associated function that needs to be used as RwLockWriteGuard::downgrade_map(..). A method would interfere with methods of the same name on the contents of the locked data.

This is equivalent to a combination of asynchronous RwLockWriteGuard::map and RwLockWriteGuard::downgrade from the parking_lot crate.

Inside of f, you retain exclusive access to the data, despite only being given a &T. Handing out a &mut T would result in unsoundness, as you could use interior mutability.

Examples

use tokio::sync::{RwLock, RwLockWriteGuard};

#[derive(Debug, Clone, Copy, PartialEq, Eq)]
struct Foo(u32);

let lock = RwLock::new(Foo(1));

let mapped = RwLockWriteGuard::downgrade_map(lock.write().await, |f| &f.0);
let foo = lock.read().await;
assert_eq!(foo.0, *mapped);

pub fn try_map<F, U>( this: RwLockWriteGuard<'a, T>, f: F, ) -> Result<RwLockMappedWriteGuard<'a, U>, RwLockWriteGuard<'a, T>>
where F: FnOnce(&mut T) -> Option<&mut U>, U: ?Sized,

Attempts to make a new RwLockMappedWriteGuard for a component of the locked data. The original guard is returned if the closure returns None.

This operation cannot fail as the RwLockWriteGuard passed in already locked the data.

This is an associated function that needs to be used as RwLockWriteGuard::try_map(...). A method would interfere with methods of the same name on the contents of the locked data.

This is an asynchronous version of RwLockWriteGuard::try_map from the parking_lot crate.

Examples

use tokio::sync::{RwLock, RwLockWriteGuard};

#[derive(Debug, Clone, Copy, PartialEq, Eq)]
struct Foo(u32);

let lock = RwLock::new(Foo(1));

{
    let guard = lock.write().await;
    let mut guard = RwLockWriteGuard::try_map(guard, |f| Some(&mut f.0)).expect("should not fail");
    *guard = 2;
}

assert_eq!(Foo(2), *lock.read().await);

pub fn try_downgrade_map<F, U>( this: RwLockWriteGuard<'a, T>, f: F, ) -> Result<RwLockReadGuard<'a, U>, RwLockWriteGuard<'a, T>>
where F: FnOnce(&T) -> Option<&U>, U: ?Sized,

Attempts to make a new RwLockReadGuard for a component of the locked data. The original guard is returned if the closure returns None.

This operation cannot fail as the RwLockWriteGuard passed in already locked the data.

This is an associated function that needs to be used as RwLockWriteGuard::try_downgrade_map(...). A method would interfere with methods of the same name on the contents of the locked data.

This is equivalent to a combination of asynchronous RwLockWriteGuard::try_map and RwLockWriteGuard::downgrade from the parking_lot crate.

Inside of f, you retain exclusive access to the data, despite only being given a &T. Handing out a &mut T would result in unsoundness, as you could use interior mutability.

If this function returns Err(...), the lock is never unlocked nor downgraded.

Examples

use tokio::sync::{RwLock, RwLockWriteGuard};

#[derive(Debug, Clone, Copy, PartialEq, Eq)]
struct Foo(u32);

let lock = RwLock::new(Foo(1));

let guard = RwLockWriteGuard::try_downgrade_map(lock.write().await, |f| Some(&f.0)).expect("should not fail");
let foo = lock.read().await;
assert_eq!(foo.0, *guard);

pub fn into_mapped( this: RwLockWriteGuard<'a, T>, ) -> RwLockMappedWriteGuard<'a, T>

Converts this RwLockWriteGuard into an RwLockMappedWriteGuard. This method can be used to store a non-mapped guard in a struct field that expects a mapped guard.

This is equivalent to calling RwLockWriteGuard::map(guard, |me| me).

pub fn downgrade(self) -> RwLockReadGuard<'a, T>

Atomically downgrades a write lock into a read lock without allowing any writers to take exclusive access of the lock in the meantime.

Note: This won’t necessarily allow any additional readers to acquire locks, since RwLock is fair and it is possible that a writer is next in line.

Returns an RAII guard which will drop this read access of the RwLock when dropped.

Examples

let lock = Arc::new(RwLock::new(1));

let n = lock.write().await;

let cloned_lock = lock.clone();
let handle = tokio::spawn(async move {
    *cloned_lock.write().await = 2;
});

let n = n.downgrade();
assert_eq!(*n, 1, "downgrade is atomic");

drop(n);
handle.await.unwrap();
assert_eq!(*lock.read().await, 2, "second writer obtained write lock");